+++ CHANGELOG.md
@@ -5,3 +5,4 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Prevent multiple iframes being loaded
+- Add `dispose()` on `FramedClient`